# k8s/src/appstore-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AppStore API",
    "description": "App catalogue, deployments and teams for the DHBW cloud self-service platform",
    "version": "0.1.0-test.1"
  },
  "paths": {
...
    "/deployments/{deployment_id}/resources": {
      "get": {
        "tags": [
          "Deployments"
        ],
        "summary": "List Deployment Resources",
        "description": "List the deployment's OpenStack resources for the Infrastructure tab.\n\nParsed from the OpenTofu state; with ``refresh=true`` (default) compute\ninstances are overlaid with live OpenStack status, fetched with the\ncaller's own credential for the project (skipped without one).\nRequires the owner view (owner, project peers, admins); 404/403\notherwise. ``refresh=false`` skips the live OpenStack join — use\nthat when polling rapidly to avoid hammering Keystone, or when\nOpenStack is known unavailable and the cached state is good enough.",
        "operationId": "listDeploymentResources",
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deployment Id"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResourceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/deployments/{deployment_id}/resources/{address}": {
      "get": {
        "tags": [
          "Deployments"
        ],
        "summary": "Get Deployment Resource Detail",
        "description": "Get the live detail (ports, security groups, volumes, metadata) of one compute instance.\n\nRequires the owner view. ``address`` is the OpenTofu state address, e.g.\n``openstack_compute_instance_v2.team_ide[\"Team-A\"]``; 422\n``invalid_resource_address`` for a malformed one, 404\n``resource_not_in_state`` when the state has no such instance.\n\nUses a ``path``-converter on the address so the for_each-key\nquoting (``team_ide[\"Team-A\"]``) survives URL routing without\naggressive encoding gymnastics on the client side. The address\nMUST exist in the cached state and MUST be a compute instance.",
        "operationId": "getDeploymentResourceDetail",
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deployment Id"
            }
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResourceSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/openstack-credentials/{credential_id}/resources/refresh": {
      "post": {
        "tags": [
          "OpenStack Resources"
        ],
        "summary": "Refresh Resource Cache",
        "description": "Drop the cached resource lists of this credential (204).\n\nTriggered by the refresh button next to a wizard picker — the user has\njust created a resource in Horizon and wants to see it. With ``kind``\nonly that kind is dropped (the ``@openstack:<kind>`` marker names, e.g.\n``network``, ``security_group``, ``availability_zone``); 422 for an\nunknown one. 404 for a credential that is not the caller's.",
        "operationId": "refreshResourceCache",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only invalidate this kind: a marker kind such as network or security_group",
              "title": "Kind"
            },
            "description": "Only invalidate this kind: a marker kind such as network or security_group"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/openstack-credentials/{credential_id}/resources/networks": {
      "get": {
        "tags": [
          "OpenStack Resources"
        ],
        "summary": "List Networks",
        "description": "List all networks visible to the credential's project.\n\n``shared`` and ``external`` are included so the frontend can render\n\"External Network\" hints. Like every endpoint here: 404 for a credential\nthat is not the caller's, 502 when OpenStack fails.",
        "operationId": "listNetworks",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkItem"
                  },
                  "title": "Response Listnetworks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/openstack-credentials/{credential_id}/resources/subnets": {
      "get": {
        "tags": [
          "OpenStack Resources"
        ],
        "summary": "List Subnets",
        "description": "List subnets, optionally only those of ``network_id``.\n\nThe filter is passed to the OpenStack API, and each filter value is\ncached under its own key next to the unfiltered list.",
        "operationId": "listSubnets",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          },
          {
            "name": "network_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only subnets of this network",
              "title": "Network Id"
            },
            "description": "Only subnets of this network"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubnetItem"
                  },
                  "title": "Response Listsubnets"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/openstack-credentials/{credential_id}/resources/flavors": {
      "get": {
        "tags": [
          "OpenStack Resources"
        ],
        "summary": "List Flavors",
        "description": "List compute flavors with vCPUs, RAM (MB) and disk (GB).\n\nPrivate flavors (``is_public=false``) are included; the frontend can\nrender a note.",
        "operationId": "listFlavors",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlavorItem"
                  },
                  "title": "Response Listflavors"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/openstack-credentials/{credential_id}/resources/images": {
      "get": {
        "tags": [
          "OpenStack Resources"
        ],
        "summary": "List Images",
        "description": "List Glance images visible to the project, by default only ``active`` ones.\n\nDefault ``status=active`` keeps \"queued\" or \"deleted\" images out of the\npicker; ``status=all`` returns every image. ``size`` is in bytes.",
        "operationId": "listImages",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Image status to list (default: active); 'all' for every status",
              "default": "active",
              "title": "Status"
            },
            "description": "Image status to list (default: active); 'all' for every status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageItem"
                  },
                  "title": "Response Listimages"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/openstack-credentials/{credential_id}/resources/keypairs": {
      "get": {
        "tags": [
          "OpenStack Resources"
        ],
        "summary": "List Keypairs",
        "description": "List the SSH keypairs of the credential's OpenStack user.\n\nNova keypairs belong to the user, not the project, and Terraform modules\nreference them by ``name``; ``id`` therefore repeats the name.",
        "operationId": "listKeypairs",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credential Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeypairItem"
                  },
                  "title": "Response Listkeypairs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeploymentResourceListResponse": {
        "properties": {
          "resources": {
            "items": {
              "$ref": "#/components/schemas/DeploymentResourceSchema"
            },
            "type": "array",
            "title": "Resources"
          },
          "live": {
            "type": "boolean",
            "title": "Live"
          }
        },
        "type": "object",
        "required": [
          "resources",
          "live"
        ],
        "title": "DeploymentResourceListResponse",
        "description": "The resources of a deployment's OpenTofu state, optionally joined with live OpenStack data.\n\nAn object rather than a bare list so metadata can be added later without a breaking change."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NetworkItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "shared": {
            "type": "boolean",
            "title": "Shared",
            "default": false
          },
          "external": {
            "type": "boolean",
            "title": "External",
            "default": false
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          }
        },
        "type": "object",
        "title": "NetworkItem",
        "description": "A Neutron network."
      },
...
    }
  }
}